[REFACTOR] Move the utils directory under src directory
